tests now using common logic
